When reading a 2-byte word from memory, wait a (very) short time before continuing to the second byte. This gives the data and address LEDs some time to show the first byte, fixing some issues with incorrect LED displays while running.
For example, the example in Mike's video #2 which is jumps to 020 and back
to 000 was not showing the D4 LED (and A0 only very dimly). With this change,
the D4 and A0 will show properly.

Inserting this delay does come with a small performance penalty but it
seems small enough to justify the change.